requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
numpy
torch>=1.5.0+cpu 
tqdm
pyyaml
numba>=0.39.0
torch_scatter>=1.3.1+cpu
Cython
scipy
dropblock
plyfile
open3d